Rename NewDigitalObjectInterface to DigitalObjectInterface
