+++ app/javascript/hyacinth_ui_v1/components/digital_objects/DigitalObjectInterface.js
