+++ app/javascript/hyacinth_ui_v1/components/digital_objects/children/Children.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import produce from 'immer';
 import { useQuery } from '@apollo/react-hooks';
 
-import DigitalObjectInterface from '../NewDigitalObjectInterface';
+import DigitalObjectInterface from '../DigitalObjectInterface';
 import TabHeading from '../../ui/tabs/TabHeading';
 import { getChildrenDigitalObjectQuery } from '../../../graphql/digitalObjects';
 import GraphQLErrors from '../../ui/GraphQLErrors';
